website/requirements.txt
bcrypt==3.1.4
cffi==1.11.5
config==0.4.0
dataclasses==0.6
Django==2.1.3
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
djongo==1.2.30
jsonpickle==1.0
mongoengine==0.8.0
psycopg2==2.7.4
pycparser==2.19
PyJWT==1.6.4
pymongo==3.7.2
pytz==2018.7
six==1.11.0
sqlparse==0.2.4
